refactor kernel process utilities
