+++ core/kernel/index.ts
@@ -19,43 +19,24 @@ import { TCP } from '../net/tcp';
 import { UDP } from '../net/udp';
 import { BASH_SOURCE } from '../fs/bin';
 import { startHttpd, startSshd, startPingService } from '../services';
-
-type ProcessID = number;
-type FileDescriptor = number;
-
-/**
- * Represents a single entry in a process's file descriptor table.
- */
-interface FileDescriptorEntry {
-  path: string;
-  position: number;
-  flags: string; // e.g., 'r', 'w', 'rw'
-  virtual?: boolean;
-}
-
-/**
- * Process Control Block: Stores the state of a single process.
- */
-interface ProcessControlBlock {
-  pid: ProcessID;
-  isolateId: number;
-  uid: number;
-  gid: number;
-  quotaMs: number;
-  quotaMs_total: number;
-  quotaMem: number;
-    cpuMs: number;
-    memBytes: number;
-    tty?: string;
-  started: boolean;
-  allowedSyscalls?: Set<string>;
-  fds: Map<FileDescriptor, FileDescriptorEntry>;
-  nextFd: FileDescriptor;
-  code?: string;
-  argv?: string[];
-  exited?: boolean;
-  exitCode?: number;
-}
+import {
+    ProcessID,
+    FileDescriptor,
+    FileDescriptorEntry,
+    ProcessControlBlock,
+    dispatcherMap,
+    createProcess,
+    cleanupProcess,
+    ensureProcRoot,
+    registerProc,
+    registerProcFd,
+    removeProcFd,
+    procStatus,
+    runProcess,
+    registerJob,
+    removeJob,
+    updateJobStatus
+} from './process';
 
 type Program = {
   main: (syscall: SyscallDispatcher, argv: string[]) => Promise<number>;
@@ -99,7 +80,6 @@ export interface Snapshot {
  */
 export type SyscallDispatcher = (call: string, ...args: any[]) => Promise<any>;
 
-const dispatcherMap: Map<ProcessID, SyscallDispatcher> = new Map();
 
 /**
  * The Helios-OS Kernel, responsible for process, file, and system management.
@@ -124,6 +104,17 @@ export class Kernel {
   private networkingStarted = false;
   private jobs: Map<number, { id: number; pids: number[]; command: string; status: string }> = new Map();
   private nextJob = 1;
+  private createProcess = createProcess;
+  private cleanupProcess = cleanupProcess;
+  private ensureProcRoot = ensureProcRoot;
+  private registerProc = registerProc;
+  private registerProcFd = registerProcFd;
+  private removeProcFd = removeProcFd;
+  private procStatus = procStatus;
+  private runProcess = runProcess;
+  public registerJob = registerJob;
+  public removeJob = removeJob;
+  public updateJobStatus = updateJobStatus;
 
   private constructor(fs: AsyncFileSystem) {
     this.state = {
@@ -335,87 +326,6 @@ export class Kernel {
       this.state.nics.set(n.id, n);
     }
     this.pendingNics = null;
-  }
-
-  private createProcess(): ProcessID {
-    const pid = this.state.nextPid++;
-    const pcb: ProcessControlBlock = {
-        pid,
-        isolateId: pid,
-        uid: 1000,
-        gid: 1000,
-        quotaMs: 10,
-        quotaMs_total: Infinity,
-        quotaMem: 8 * 1024 * 1024,
-        cpuMs: 0,
-        memBytes: 0,
-        tty: undefined,
-        started: false,
-        allowedSyscalls: undefined,
-        fds: new Map(),
-        nextFd: 3, // 0, 1, 2 are reserved for stdio
-        exited: false,
-    };
-    const processes = new Map(this.state.processes);
-    processes.set(pid, pcb);
-    this.state = { ...this.state, processes };
-    this.registerProc(pid);
-    return pid;
-  }
-
-  private cleanupProcess(pid: ProcessID) {
-    const processes = new Map(this.state.processes);
-    processes.delete(pid);
-    this.state = { ...this.state, processes };
-  }
-
-  private ensureProcRoot() {
-    if (!(this.state.fs as any).getNode('/proc')) {
-      (this.state.fs as any).createVirtualDirectory('/proc', 0o555);
-    }
-  }
-
-  private registerProc(pid: ProcessID) {
-    this.ensureProcRoot();
-    if (!(this.state.fs as any).getNode(`/proc/${pid}`)) {
-      (this.state.fs as any).createVirtualDirectory(`/proc/${pid}`, 0o555);
-    }
-    if (!(this.state.fs as any).getNode(`/proc/${pid}/status`)) {
-      (this.state.fs as any).createVirtualFile(`/proc/${pid}/status`, () => this.procStatus(pid), 0o444);
-    }
-    if (!(this.state.fs as any).getNode(`/proc/${pid}/fd`)) {
-      (this.state.fs as any).createVirtualDirectory(`/proc/${pid}/fd`, 0o555);
-    }
-  }
-
-  private registerProcFd(pid: ProcessID, fd: number) {
-    const pcb = this.state.processes.get(pid);
-    if (!pcb) return;
-    if (!(this.state.fs as any).getNode(`/proc/${pid}/fd/${fd}`)) {
-      (this.state.fs as any).createVirtualFile(`/proc/${pid}/fd/${fd}`, () => {
-        const entry = pcb.fds.get(fd);
-        return new TextEncoder().encode(entry ? entry.path : '');
-      }, 0o444);
-    }
-  }
-
-  private removeProcFd(pid: ProcessID, fd: number) {
-    const path = `/proc/${pid}/fd/${fd}`;
-    if ((this.state.fs as any).getNode(path)) {
-      (this.state.fs as any).remove(path);
-    }
-  }
-
-  private procStatus(pid: ProcessID): Uint8Array {
-    const pcb = this.state.processes.get(pid);
-    if (!pcb) return new Uint8Array();
-    const enc = new TextEncoder();
-    const cmd = pcb.argv ? pcb.argv.join(' ') : '';
-    const out =
-      `pid:\t${pid}\nuid:\t${pcb.uid}\n` +
-      `cpuMs:\t${pcb.cpuMs}\nmemBytes:\t${pcb.memBytes}\n` +
-      `tty:\t${pcb.tty ?? ''}\ncmd:\t${cmd}\n`;
-    return enc.encode(out);
   }
 
   private createSyscallDispatcher(pid: ProcessID): SyscallDispatcher {
@@ -718,28 +628,6 @@ export class Kernel {
     return Array.from(this.jobs.values());
   }
 
-  public registerJob(pids: number[], command: string): number {
-    const id = this.nextJob++;
-    const jobs = new Map(this.jobs);
-    const entry = { id, pids, command, status: 'Running' };
-    jobs.set(id, entry);
-    this.jobs = jobs;
-    return id;
-  }
-
-  public removeJob(id: number): void {
-    const jobs = new Map(this.jobs);
-    jobs.delete(id);
-    this.jobs = jobs;
-  }
-
-  public updateJobStatus(id: number, status: string): void {
-    const job = this.jobs.get(id);
-    if (!job) return;
-    const jobs = new Map(this.jobs);
-    jobs.set(id, { ...job, status });
-    this.jobs = jobs;
-  }
 
   public snapshot(): Snapshot {
     const replacer = (_: string, value: any) => {
@@ -802,51 +690,6 @@ export class Kernel {
     return JSON.parse(JSON.stringify(state, replacer));
   }
 
-  private async runProcess(pcb: ProcessControlBlock): Promise<void> {
-    if (!pcb.started && !pcb.code) return;
-    const syscall = this.createSyscallDispatcher(pcb.pid);
-    dispatcherMap.set(pcb.pid, syscall);
-    const args: Record<string, any> = {
-        pid: pcb.isolateId,
-        sliceMs: pcb.quotaMs,
-        quotaMem: pcb.quotaMem,
-    };
-    if (!pcb.started) {
-        const wrapped = `const main = ${pcb.code}; main(syscall, ${JSON.stringify(pcb.argv ?? [])});`;
-        args.code = wrapped;
-    }
-    try {
-        const result: any = await invoke('run_isolate_slice', args);
-        if (!pcb.started) {
-            pcb.started = true;
-            pcb.code = undefined;
-        }
-        if (result) {
-            pcb.cpuMs += result.cpu_ms ?? 0;
-            pcb.memBytes += result.mem_bytes ?? 0;
-            if (pcb.cpuMs > pcb.quotaMs_total || pcb.memBytes > pcb.quotaMem) {
-                console.warn('Process', pcb.pid, 'exceeded quota');
-                this.syscall_kill(pcb.pid, 9);
-            } else if (!result.running) {
-                pcb.exitCode = result.exit_code ?? 0;
-                pcb.exited = true;
-            }
-        } else {
-            pcb.exitCode = 0;
-            pcb.exited = true;
-        }
-    } catch (e) {
-      console.error('Process', pcb.pid, 'crashed or exceeded quota:', e);
-      pcb.exitCode = 1;
-      pcb.exited = true;
-    }
-    if (pcb.exited) {
-      try {
-        await invoke('drop_isolate', { pid: pcb.isolateId });
-      } catch {}
-    }
-    dispatcherMap.delete(pcb.pid);
-  }
 
   public async start(): Promise<void> {
     this.running = true;
